feat: All of my work for the past year
This is somewhat messy (try forgetting to commit for an entire year)
but here's a relatively complete list of changes:
- Rewrote pre-commit config from scratch (with the same checkers) and
  ensured the checks work with the latest versions (on Fedora) of each
  tool, in particular e.g. clang-tidy/cppcheck/cpplint, and documented
  more throughly the configs for each of those tools
- Added 23 bugs I've reported/fixed to BUGS-FOUND.txt
- Renamed YALIBCT_DISABLE_ALL_POSSIBLE_TESTS_BY_DEFAULT to
  YALIBCT_DISABLE_ALL_POSSIBLE_TESTS (renaming the others will be in
  the next commit)
- Removed
  YALIBCT_DISABLE_PRINTF_FIELD_LENGTH_ON_S_CONVERSION_SPECIFIER_TESTS
  in favor of
  YALIBCT_DISABLE_PRINTF_FIELD_WIDTH_ON_WIDE_S_CONVERSION_SPECIFIER_TESTS
- Added YALIBCT_DISABLE_PRINTF_I_CONVERSION_SPECIFIER_TESTS,
  YALIBCT_DISABLE_PRINTF_PRECISION_ON_WIDE_S_CONVERSION_SPECIFIER_TESTS,
  YALIBCT_DISABLE_PRINTF_PRECISION_ON_WIDE_C_CONVERSION_SPECIFIER_TESTS,
  YALIBCT_DISABLE_PRINTF_FIELD_WIDTH_TESTS,
  YALIBCT_DISABLE_PRINTF_FIELD_WIDTH_ON_WIDE_S_CONVERSION_SPECIFIER_TESTS,
  YALIBCT_DISABLE_PRINTF_FIELD_WIDTH_ON_WIDE_C_CONVERSION_SPECIFIER_TESTS,
  YALIBCT_DISABLE_SNPRINTF_SIZE_LARGER_THAN_ACTUAL_BUFFER_TESTS,
  YALIBCT_DISABLE_ALL_NON_STANDARD_TESTS
- Added _DARWIN_C_SOURCE, __EXTENSIONS__, _OPENBSD_SOURCE,
  _POSIX_SOURCE, _POSIX_1_SOURCE=2, _POSIX_PTHREAD_SEMANTICS,
  __STDC_WANT_IEC_60559_ATTRIBS_EXT__,
  __STDC_WANT_IEC_60559_BFP_EXT__, __STDC_WANT_IEC_60559_DFP_EXT__,
  __STDC_WANT_IEC_60559_EXT__, __STDC_WANT_IEC_60559_FUNCS_EXT__,
  __STDC_WANT_IEC_60559_TYPES_EXT__, __STDC_WANT_LIB_EXT2__,
  __STDC_WANT_MATH_SPEC_FUNCS__, _TANDEM_SOURCE and _COSMO_SOURCE to
  the list of defines we use to try and enable as many features as
  possible in the libc we're testing
- Moved around a few functions in CMakeLists.txt
- Added printf and fstatat to the list of functions we test for
- Tweaked a few comments in CMakeLists.txt
- Noted that the MIT license only applies to code not imported from
  other repositories
- Added a bunch of notes to TODO
- Fixed a bunch of files not including the headers they need
- Changed a few variables/parameters to be const where possible
- Added names for parameters to some functions that lacked them
- Initialized some variables that clang-tidy thought could be used
  uninitialized
- Added/removed some casts where clang-tidy thought I should do so
- Changed an odd if-statement in cmocka.h to be slightly less
  redundant
- Added a bunch of NOLINT statements in places where clang-tidy was
  saying stupid stuff
- Fixed a few printf statements to take the correct types as input
- Fixed a memory leak of `optionstr` in cygwin.h
- Fixed indentation a bit in places where clang-tidy thought I should
  do so
- Added some asserts where judicious to make some things that are
  always true clearer to clang-tidy
- Refactored some identical switch cases
- Added some default cases on a few switches
- Commented out 1 or 2 unused variables
- Added YALIBCT_DIAGNOSTIC_IGNORE_WFORMAT_TRUNCATION and used it as
  appropriate
- Fixed size of bytes array in portable-snippets-endian.h
- Made a macro into an enum in a few places where clang-tidy thought
  it appropriate
- Cast a few function calls to `(void)` to make it clear they're meant
  to have their result ignored
- Added include/test-lib/portable-symbols/printf.h
- Change STATIC_YOINK to __static_yoink for cosmopolitan since they
  changed the name upstream. Also minimized the amount of times we use
  it to what's necessary to get a backtrace on a crash
- Changed the name of some of the build directories on the scripts in
  internal/devel-scripts
- Rewrote each script in internal/devel-scripts and made it so they
  document exactly why each of the flags is necessary. Also took the
  opportunity to fix a bunch of bugs in cosmopolitan, dietlibc and
  musl
- Added a script for cosmopolitan latest release rather than latest
  git
- Added build-gcc-x86_64-redhat-linux-glibc-without-many-functions.sh
  script (to be filled in later)
- Fixed a bit of formatting/comments/messages in run_test_binaries.sh
- Made it so that run_test_binaries.sh handles a few more hexadecimal
  floats for the littlekernel tests
- Made it so that the jobs count is updated on each loop in the job
  handling code in run_test_binaries.sh

+++ include/test-lib/portable-symbols/zerosize_ptr.h
@@ -74,7 +74,7 @@ zerosize_ptr (void)
     {
       int pagesize = getpagesize ();
       char *two_pages =
-        (char *) mmap (NULL, 2 * pagesize, PROT_READ | PROT_WRITE,
+	(char *) mmap (NULL, 2 * (size_t)pagesize, PROT_READ | PROT_WRITE,
                        flags, fd, 0);
       if (two_pages != (char *)(-1)
           && mprotect (two_pages + pagesize, pagesize, PROT_NONE) == 0)
